Drop support for ASDF2, gain support for future ASDF 3.1.8 or so
Use perform-lisp-compilation and drop support for ASDF2.
asdf-support was full of things only meant to support the long gone ASDF2,
including a bad old copy of the perform method for compile-op, instead of
using perform-lisp-compilation as should be used in ASDF3. This probably
already made cl-protobufs not work in some corner cases, but would have made it
stop working at all when compile-op-flags will be dropped from ASDF in 3.1.x
(for x > 7).

Also minor cleanups of asd files.